+++ lending/src/sdk/models/operations/getaccountingdirectcost.ts
@@ -7,14 +7,20 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetAccountingDirectCostRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
     /**
-     * Unique identifier for a direct cost
+     * Unique identifier for a direct cost.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=directCostId" })
     directCostId: string;
@@ -27,6 +33,9 @@ export class GetAccountingDirectCostResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     accountingDirectCost?: shared.AccountingDirectCost;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class GetAccountingDirectCostResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }